Pushing major changes and redo assignment 2
Using the new data given, I renamed all of the columns to match the meta data to run PCA plot file. We are missing a sample now so i removed it from the gse_matrix. The new file for the data is GoatDataModified.tsv and the edited meta data is gse_matrix2. Our data now has 53 samples instead of 54. C_27 had to be removed.